refactor: refactor folder structure
