+++ src/components/EQ.tsx
