+++ src/types/game.d.ts
@@ -1,10 +1,8 @@
-
-
 export type GameMemoType = {
-    score: number
-    strike: number
-    mistakes: number
-    userId: string
-    id: string
+  score: number
+  strike: number
+  mistakes: number
+  userId: string
+  level: number
+  id: string
 }
-
